+++ src/pyramid_bimt/tests/test_user_model.py
@@ -12,6 +12,7 @@ from pyramid_bimt.testing import initTestingDB
 from pyramid_bimt.tests.test_group_model import _make_group
 from sqlalchemy.exc import IntegrityError
 
+import mock
 import unittest
 
 
@@ -406,6 +407,16 @@ class TestUserProperties(unittest.TestCase):
 
         self.user.set_property('foo', u'baz')
         self.assertEqual(self.user.get_property('foo'), u'baz')
+
+    @mock.patch('pyramid_bimt.security.get_current_registry')
+    def test_set_property_secure(self, get_current_registry):
+        get_current_registry.return_value.settings = {
+            'bimt.encryption_aes_16b_key': 'abcdabcdabcdabcd',
+        }
+        self.user.set_property('foo', u'bar', secure=True)
+        self.assertEqual(self.user.get_property('foo', secure=True), u'bar')
+        self.assertNotEqual(
+            self.user.get_property('foo', secure=False), u'bar')
 
     def test_set_property_strict(self):
         with self.assertRaises(KeyError) as cm: